requirements.txt.txt

numpy==1.24.3
pandas
boto3
Pillow==10.0.0
PyYAML==6.0.1
torch ==2.0.0
torchvision
ultralytics
opencv-python-headless==4.8.0.74

